W-12622230: Re do test here
